+++ app/src/main/java/com/lozog/expensetracker/util/SheetsStatus.kt
@@ -1,4 +1,4 @@
-package com.lozog.expensetracker
+package com.lozog.expensetracker.util
 
 enum class SheetsStatus {
     DONE, IN_PROGRESS
