+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_form"
-        android:name="com.lozog.expensetracker.ui.FormFragment"
+        android:name="com.lozog.expensetracker.ui.form.FormFragment"
         android:label="Expense Tracker"
         tools:layout="@layout/fragment_form" />
 
     <fragment
         android:id="@+id/navigation_settings"
-        android:name="com.lozog.expensetracker.ui.SettingsFragment"
+        android:name="com.lozog.expensetracker.ui.settings.SettingsFragment"
         android:label="Settings"
         tools:layout="@layout/fragment_settings"/>
 
     <fragment
         android:id="@+id/navigation_account"
-        android:name="com.lozog.expensetracker.ui.AccountFragment"
+        android:name="com.lozog.expensetracker.ui.account.AccountFragment"
         android:label="Account"
         tools:layout="@layout/fragment_account"/>
